fixing vhh inference and adding pretrain loader
