+++ src/ibex/__init__.py
@@ -1,2 +1,2 @@
 from .model import Ibex
-from .predict import MODEL_CHECKPOINTS, ENSEMBLE_MODELS, checkpoint_path, inference, batch_inference
+from .predict import inference, batch_inference
